Create Input Section 2 - Income Table  (#337)
* 1st version

* 1st commit

* Simple Grid

* storybook render

* Basic Expense Table Hardcoded

* Table styled to Figma + input fields

Need to figure out states next to save the value and sum A+B and validate against C boolean

* Got Input Fields on same line as Field Letters (A,B,C)

* Update IncomeTable.tsx

* State variable passes

* States passed with Sum check (print)

* Updated the props check in pages tsx

* ternary attempt to checksum

probably not the right way but curious to know why the logic isn't working.

* Sum Validation - with ternary operators

* Validation

Ran into an issue however on the testing. Otherwise helper function seems to work.

* Colour Cell Works!

Need to try and clean up.

* Works

but tunnel visioned and went back to a poor code version. Got lost and ended up with a better version of my bad first attempt lol

* Rendering Works

* Update IncomeTable.tsx