Smooth Scrolling (Works-sorta)
Works for the first 2 headers but not last as well. i think it has to do with if the section is already in view partly then it wont scroll there.

Im going to try and space the table sections out each to a full screen and see if this resolves it