Finance smooth scrolling (#426)
* 1st version

* Revert "Merge branch 'main' of https://github.com/vithushan19/math into main"

* Smooth Scrolling (Works-sorta)

Works for the first 2 headers but not last as well. i think it has to do with if the section is already in view partly then it wont scroll there.

Im going to try and space the table sections out each to a full screen and see if this resolves it

* Smooth Scroll Fully Working!

Done after this review.
Thing to note..sometimes it works...scrolling up and down there's a delay or sometimes it doesnt auto scroll. But nothing critical...should be covered in more testing.