Finance badge (#437)
* 1st version

* Setting Up for Badge

Added Finance Grade - but later will have grades 1-6 be switches for the different amounts saved. Want this exercise to evolve to an exercise in having the highest monthly savings.

* corrected review and added badge image

* added finance as.a topic

* FetchBadge ID based on User ID Query

* testing pulling badges on ID

created a new page to test on. will merge with Danny's stage once I lock this down. [not working thouugh]

* query update

* get the badges

* Badge check

* working through the DB badge pull

* trying to pass badges as a prop

unsuccessful attempt

* helper function error

{data...} isn't a string...thought I could define my prop as an array...not working

* badgeImage error in the first stage setup

* Passing Data from parent to RulesSession

With a print screen to see if it shows up (which it does not)

* Data doesn't seem to be passing

* It workrs!

next up grphQL filtering down to just the IDS I want (addition/subtraction)

* Working Badge (mul/Div)

* user login

* Write Finance badge for USER ID

* Updated

* next iteration

* Helper function not working

* update - helper function working

BUT my lines 31/32 to show the badges isn't. Not coming out on the page

* badges not rendering outside of Div

* Working!

Gradient now set for when addition and subtraction not found!

If this passes - final stage is just to correct for login and actual sub/addition.

* Vithusan's comments addressed

all issues addressed
Hardcoded userID removed....should be good to merge

* Update finance-profile.tsx

* Final Review and Merge

* Update RulesSession.tsx

* Final comments

Co-authored-by: vithushan19 <vithushan19@gmail.com>