improve CalculateWinner component
Move calculations of image and winner to the main game page due to not being able to move the gameState variable definition as an import and to also make it cleaner.

In doing so, added the props of winner and image to the CalculateWinner component.

Also updated the path for the images so that they appear.

finally added numberOfToys to the SalesTaxModel, as it was not there.