Enhance UI for CALCULATE_WINNER stage (#753)
* working on layout for the grid of blocks

working on making the grid layout work

* adding a for loop to separate blocks

trying to make the UI layout easier

* wrap the mapped gameState function with divs to create individual boxes for each block

and also used slice on the gameState array to put the boxes in the right order

* justify-between the buttons

working on the layout of the title/game board

* create opening page to multiplication block game

add UI and content components

* create UI for CALCULATE_WINNER stage

Added images/text depending on the outcome of the game. Enhanced UI.

* fixed the index

adjusted the index appropriately

* move images into longestStreak folder

* update test file

* create components SetRules and CalculateWinner

put these components in the longestStreak folder in the Math1 folder

Also moved all the functions in Game outside of the default function in order to allow them to be exported later if needed.

* create a random number generator

working to create a function that will map to the text key of the object gameState that will generate a pair of strings that will be "x*y" and its product.  Those will be the values showing on the blocks.

* fixed randomNumberProductList

removed the reduce function and used a for loop to fill the initialGameState object which will be called later on in the code to fill the blocks.

* call randomNumberProductList to put numbers on blocks

also update product string to print random numbers

* improve CalculateWinner component

Move calculations of image and winner to the main game page due to not being able to move the gameState variable definition as an import and to also make it cleaner.

In doing so, added the props of winner and image to the CalculateWinner component.

Also updated the path for the images so that they appear.

finally added numberOfToys to the SalesTaxModel, as it was not there.

* simplified code for resetting game

removed handleReset and simply used handlePlayGame which included resetting the game board with numbers.

* fix slice and index on game page

adjust the slicing and the index and adding in two reverse mappings on the gameState2 array so that the blocks are ordered like a snake in one continuous line around the edge.