beginning slice file (#760)
* working on layout for the grid of blocks

working on making the grid layout work

* adding a for loop to separate blocks

trying to make the UI layout easier

* wrap the mapped gameState function with divs to create individual boxes for each block

and also used slice on the gameState array to put the boxes in the right order

* justify-between the buttons

working on the layout of the title/game board

* create opening page to multiplication block game

add UI and content components

* create UI for CALCULATE_WINNER stage

Added images/text depending on the outcome of the game. Enhanced UI.

* fixed the index

adjusted the index appropriately

* move images into longestStreak folder

* update test file

* create components SetRules and CalculateWinner

put these components in the longestStreak folder in the Math1 folder

Also moved all the functions in Game outside of the default function in order to allow them to be exported later if needed.

* create a random number generator

working to create a function that will map to the text key of the object gameState that will generate a pair of strings that will be "x*y" and its product.  Those will be the values showing on the blocks.

* fixed randomNumberProductList

removed the reduce function and used a for loop to fill the initialGameState object which will be called later on in the code to fill the blocks.

* call randomNumberProductList to put numbers on blocks

also update product string to print random numbers

* improve CalculateWinner component

Move calculations of image and winner to the main game page due to not being able to move the gameState variable definition as an import and to also make it cleaner.

In doing so, added the props of winner and image to the CalculateWinner component.

Also updated the path for the images so that they appear.

finally added numberOfToys to the SalesTaxModel, as it was not there.

* simplified code for resetting game

removed handleReset and simply used handlePlayGame which included resetting the game board with numbers.

* fix slice and index on game page

adjust the slicing and the index and adding in two reverse mappings on the gameState2 array so that the blocks are ordered like a snake in one continuous line around the edge.

* add tailwind coding to fix how the input box looks

also adding the name that is chosen to the end winner page.

* begin redux reducer functions

* Update longestStreakSlice.ts

putting down some ideas for redux state

* add redux to manage state

using redux to change between the stages.

* renamed calculateWinner component

renamed component to Winner

* move functions out of default function

clean up j-test file

* updating redux for blockState

* changes requested on review

made changes requested on review

* finish redux coding

moved initalGameState into the slice folder and then changed state in the game.ts code using the redux instead of the state hooks from before.

* WIP adding AI component

working on adding the AI selection for the blocks.

* Add start of AI selection

* fix imports and tests

* add test case

* add todos

* changes for removing isPlayerOneActive

removing the last two-player components of the game and using isPlayerActive to activate the AI player through redux.

* add logic to block wrong selection

add logic to not allow player to reselect an AI selected block

* filter for second AI block selection

code finds a a non-selected block that has the value to search for and then selects that block

* update test on longestSlice

added a test for the two AI selected blocks

* add currentlySelectedBlock

trying to get this to store the index of the block so that it can be recalled after the next click.

* updated test file with selecting the correct block

* adjust test expect values for longestStreakSlice.test

* update longestStreakTest file to reflect the slice

changes were made in the slice and are now mirrored in the test

* leave second test for next branch

Co-authored-by: vithushan19 <vithushan19@gmail.com>