cleaning up win functions and mobile responsiveness (#793)
* add levels to game initialization

* use effect

moved both setLevel and the initialize game into the use effect, and asked it to update again only based on the state changing for the level

* add levels

added BEGINNER_ADVANCED and INTERMEDIATE_ADVANCED levels

* update game.test.ts

needed to reflect the newly defined enum gameLevel.BEGINNER

* changes for pr #777

capitalized the enum GameLevel
removed level as a dependency for use effect

* created a helper function

this function will simplify the initializing the game board based on the different levels

also moved extra functions from the games file to the API folder

* update test

* moving the levels logic to an object

this will enable an easier addition of properties or levels later on.
removed setLevel redux and moved all setting of levels and initiating state changes in the initializeGame.

* update on tests

adjusted them for the new level object change

* WIP querying currentLevel from graph ql

* adding useQuery and useMutation

* update useMutation hooks

removed two of the useMutation files and placed the updateGameLevel in the Winner component so that a user can save his/her progress

* add alert to show game was saved

* PR Feedback

* adjusted tests due to added functionality with game levels

* working on upsert

* PR fixes

* cleaning up the handleSaveGameClick function

* add updateGameLevel

added a function to show the winner
passed it as a prop from the game page to the child "Winner" component
moved the upsert useMutation to the beginning of the game to add new users if they are not in hasura
add updateGameLevel in the Winner component to increment the game level if the player wins

* clean up longestStreak.ts file in api folder

created a main calculateWinner function and then the associated helper functions to simplify code and allow for anything to be easily rendered based on which player wins

had to move showEndGameImage to a  javascript file in the api folder due to the fact that it returns a component and images.

* update game levels to make them easier

* update tests to match changes

* add placeholder function on input box

* cleaning up code

* add downgradeGameLevel when user 1 loses

* update welcome message

* update upgrade and downgrade mutations

 allowing them to stay in the parameters of current level limits in the code

* update UI

create a toggle button and add highlights to blocks on main board

* try refetch queries

* working on dropdown

* changes requested on merge

* mobile responsiveness and updating queries

added mobile responsive tailwind code in particular for the game board. additionally updated the query/mutations to include games played/lost/won
currently working on finalizing a table that will display user statistics.  Trying to get the query working correctly

* PR Feedback

* rework winner function code

rework winner function to make it ternary and simpler.
create an object called gameWinsMetaData that contains the winner information and pass that into the winner component from the main game page after winner is calculated

separate fetch queries into separate files.

move buttons into columns

* clean up functions and increase mobile responsiveness

* additional mobile changes

* update tests

* updates for mobile responsiveness on main board

* changes requested for PR

added type WinningOutcomes that included winnerOutcomes:WinnerData

removed tables that were being unused.

removed commented out code.

removed unused return value

* renaming file to LongestStreakGame

Co-authored-by: vithushan19 <vithushan19@gmail.com>