added a mockDate library
In order to fully test the date functions, I added a mock date library, which allowed me to set a mock date in the jest file that sets the new date to a mock date as specified in the test.

I also added a test to test if the highlight should be yellow.