added a hasRun state hook
This will track whether the component  has run or not and will therefore not trigger a re-run of the useEffect when skillRatings is updated after the page loads initially.  this will enable the useEffect to keep checking for a skillRatings variable and run multiple times before setting the sections, but will not re-run when the skillRatings object changes when the slider is moved because hasRun will be set to true.