bypassing additional query if userAssignments global state exists (#849)
* added a check to reduce number of network calls on the fetching assignments, if possible

also removed local state to set the userAssignments and used the redux global state

* updated template to check for redux state of userAssignments before query is run.

* created function deployCurrentStage

this cleans up the logic so it doesn't need to be repeated.

* updated logic in query and added assignment id

* fixing linter errors

* remove promise.all syntax

* allow for an optional loom video on the output section

1. created a ScreenshotOrVideoId property with a discriminated union type that forces the property to be one or the other, not both, but one is required.
2. Also when deployed in the AssignmentComponent had it check for a screenshot and then deployed that, if not, give the videoId

Question: would we want both at any time?  If so, they could be separate properties and made optional

* removed additional promise.all

* changed the property renders to match the new types

* fixing linter errors in AssignmentComponent

Note: the enum Stage, the linter said the variable was declared and not read. It is read in the template file, so I disabled the linter for that variable.

* console logs

* remove leader spacing

removed the special class grid-cols-resource-row because it contained the leader spacing and put a parent div around the icon and the showHint text.

* starting to add styling from Vithushan

WIP: Drop down conditional render of hints
Added a hover transition to the Hint Row to make that appear bigger when user hovers over it

* added some expandable containers

WIP maybe need to create a component for this to make code tighter

* smaller changes:

just made it a string and then upon implementation searched to see if screenshotOrVideoId.includes('.') so that if there is a png or jpg, it would render an image and if not, it would render the video

also fixed submission dropdown -- state hook

removed title from the data and made it optional

* removed lazy query

1. refactored everything to just run the query because we out that redux resets when you do server side rendering

* styling changes and validation on text input

1. all headers are set open by default except for one, the Hints section.  They all have hover states that are grey and when any of the sections are closed, they will grey out.  Hints pulses in order to call attention to it.

* removed title in AssignmentComponentData type

* linter cleaned up a few things and removed a space before the text-black-500

* moved enum Stage to the template file because it's the only place it's used.

* fixed linter indexing by adding key

* disabled button if the validateInput() fails

also added check for 0 characters and over 80

Co-authored-by: Vithushan <vithushan19@users.noreply.github.com>